Move TransactionDomainGroup  class
